handle recursive include_next
When the include path contains the Watt-32 header directory multiple
times, #include_next breaks, since two headers have it located inside
the include guard.

This scenario can happen when a project builds a local copy of Watt-32,
but the user has their own copy in the global include path already.  In
this case we want to use the local version of the header, so the
recursive #include_next must be placed at the end of the header file.